+++ src/main/java/org/cicirello/search/operators/integers/IntegerVectorInitializer.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2024 Vincent A. Cicirello
+ * Copyright (C) 2002-2026 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -40,7 +40,7 @@ import org.cicirello.search.representations.IntegerVector;
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, <a
  *     href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
-public class IntegerVectorInitializer implements Initializer<IntegerVector> {
+public final class IntegerVectorInitializer implements Initializer<IntegerVector> {
 
   private final EnhancedSplittableGenerator generator;
   private final int[] x;
@@ -224,14 +224,6 @@ public class IntegerVectorInitializer implements Initializer<IntegerVector> {
     // each instance needs their own independent instances of these
     x = new int[other.x.length];
     generator = other.generator.split();
-  }
-
-  @SuppressWarnings("deprecation")
-  @Override
-  protected final void finalize() {
-    // Prevents potential finalizer vulnerability from exceptions thrown from constructors.
-    // See:
-    // https://wiki.sei.cmu.edu/confluence/display/java/OBJ11-J.+Be+wary+of+letting+constructors+throw+exceptions
   }
 
   @Override